+++ src/main/java/com/blackrook/rookscript/struct/Utils.java
@@ -794,6 +794,31 @@ public final class Utils
 	}
 
 	/**
+	 * Returns the file's name, no extension.
+	 * @param file the file.
+	 * @param extensionSeparator the text or characters that separates file name from extension.
+	 * @return the file's name without extension.
+	 */
+	public static String getFileNameWithoutExtension(File file, String extensionSeparator)
+	{
+		return getFileNameWithoutExtension(file.getName(), extensionSeparator);
+	}
+
+	/**
+	 * Returns the file's name, no extension.
+	 * @param filename the file name.
+	 * @param extensionSeparator the text or characters that separates file name from extension.
+	 * @return the file's name without extension.
+	 */
+	public static String getFileNameWithoutExtension(String filename, String extensionSeparator)
+	{
+		int extindex = filename.lastIndexOf(extensionSeparator);
+		if (extindex >= 0)
+			return filename.substring(0, extindex);
+		return "";
+	}
+
+	/**
 	 * Puts a short into an array.
 	 * Writes 2 bytes.
 	 * @param value the value to convert.